DOSV supports more files
Extends the number of files it can show up to minimum 31. The
shorter the file names are, the more the response can show.

This becomes possible by generating the response on the fly
not just with FIB but file body as well. The file list is loaded
at FF00h~FFFFh (256 bytes), and the code displaying the list
is embedded in the FIB extension area.